@@ -1,6 +1,7 @@
-import { ConformCashPaymentDto } from "../../dto/paymentRes.dto";
-import { PaymentReq } from "../../types/request";
-import { StripeCheckoutSessionRes } from "../../types/response";
+import { IncomingHttpHeaders } from 'http';
+import { ConformCashPaymentDto } from '../../dto/paymentRes.dto';
+import { PaymentReq } from '../../types/request';
+import { StripeCheckoutSessionRes } from '../../types/response';
 
 export interface IPaymentService {
   ConfirmCashPayment(data: {
@@ -9,5 +10,4 @@ export interface IPaymentService {
     driverId: string;
     amount: number;
   }): Promise<ConformCashPaymentDto>;
-  createCheckoutSession(data: PaymentReq): Promise<StripeCheckoutSessionRes> 
 }
